requirements.txt
attrs==20.3.0
autobahn==20.12.2
Automat==20.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
constantly==15.1.0
cryptography==3.3.1
dateparser==1.0.0
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
numpy==1.19.4
pandas==1.1.5
patsy==0.5.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==20.0.0
python-binance==0.7.5
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.4
regex==2020.11.13
requests==2.25.0
scipy==1.5.4
service-identity==18.1.0
six==1.15.0
statsmodels==0.12.1
Twisted==20.3.0
txaio==20.4.1
tzlocal==2.1
unsync==1.2.1
urllib3==1.26.2
zope.interface==5.2.0
